requirements.txt
# requirements.txt

alembic==1.11.1
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
asyncpg==0.28.0
Authlib==1.2.0
backcall==0.2.0
bcrypt==4.0.1
black==23.3.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
executing==2.0.1
expects==0.9.0
fastapi==0.103.2
flake8==6.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
iniconfig==2.0.0
ipython==8.14.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
loguru==0.7.2
lxml==4.9.3
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.4.0
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.0.2
pydantic_core==2.20.1
pyflakes==3.0.1
Pygments==2.18.0
PyJWT==2.7.0
pytest==7.3.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.16
SQLAlchemy-Utils==0.41.1
stack-data==0.6.3
starlette==0.27.0
traitlets==5.14.3
types-passlib==1.7.7.12
types-requests==2.31.0.7
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.22.0
wcwidth==0.2.13
